+++ examples/plot_flux.py
@@ -18,26 +18,33 @@ km3pipe.style.use('moritz')
 
 ##############################################
 # define energy + zenith range
-n_points = 100
+n_points = 1000
 
 zen = np.linspace(0, np.pi, n_points)
 ene = np.logspace(0, 2, n_points)
 
 ##############################################
-# look at numu flux
+# look at numu flux (binned vs interpolated)
 
 numu_flux = Honda2015('nu_mu')
-print(
-    numu_flux(ene)
+plt.plot(ene,
+         numu_flux(ene, interpolate=True), label='Interpolated'
 )
+plt.plot(ene,
+         numu_flux(ene, interpolate=False), label='Binned'
+)
+plt.yscale('log')
+plt.xscale('log')
+plt.legend()
 
 ##############################################
-# Plot vs energy, for multiple flavors
+# Plot vs energy, for multiple flavors,
+# interpolated by default
 
 flavors = {'nu_mu', 'anu_mu', 'nu_e', 'anu_e'}
 for flav in flavors:
     flux = Honda2015(flavor=flav)
-    plt.plot(ene, flux(ene), label=flav)
+    plt.plot(ene, flux(ene, interpolate=True), label=flav)
 
 plt.yscale('log')
 plt.xscale('log')