+++ km3flux/flux.py
@@ -10,14 +10,17 @@ from scipy.integrate import romberg, simps
 from scipy.interpolate import splrep, splev, RectBivariateSpline
 
 from km3pipe.mc import name2pdg, pdg2name
-from km3flux.data import (HONDAFILE, dm_gc_spectrum, dm_sun_spectrum,
-                          DM_GC_FLAVORS, DM_GC_CHANNELS, DM_GC_MASSES,
-                          DM_SUN_FLAVORS, DM_SUN_CHANNELS, DM_SUN_MASSES
+from km3flux.data import (HONDAFILE, DM_GC_FLAVORS, DM_GC_CHANNELS,
+                          DM_GC_MASSES, DM_GC_FILE,
+                          #dm_gc_spectrum, dm_sun_spectrum,
+                          # DM_SUN_FLAVORS, DM_SUN_CHANNELS, DM_SUN_MASSES
                          )
 from km3pipe.tools import issorted, bincenters
 
 FLAVORS = ('nu_e', 'anu_e', 'nu_mu', 'anu_mu')
 MCTYPES = [name2pdg(name) for name in FLAVORS]
+
+logger = logging.getLogger(__name__)
 
 
 class BaseFlux(object):
@@ -50,33 +53,43 @@ class BaseFlux(object):
     def __init__(self, **kwargs):
         pass
 
-    def __call__(self, energy, zenith=None):
+    def __call__(self, energy, zenith=None, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
         energy = np.atleast_1d(energy)
+        logger.debug("Entering __call__...")
         if zenith is None:
-            return self._averaged(energy)
+            logger.debug("Zenith is none, using averaged table...")
+            return self._averaged(energy, interpolate=interpolate)
         zenith = np.atleast_1d(zenith)
         if len(zenith) != len(energy):
             raise ValueError("Zenith and energy need to have the same length.")
-        return self._with_zenith(energy=energy, zenith=zenith)
+        logger.debug("Zenith available, using angle-dependent table...")
+        return self._with_zenith(energy=energy, zenith=zenith,
+                                 interpolate=interpolate)
 
-    def _averaged(self, energy):
+    def _averaged(self, energy, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
         raise NotImplementedError
 
-    def _with_zenith(self, energy, zenith):
+    def _with_zenith(self, energy, zenith, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
         raise NotImplementedError
 
-    def integrate(self, zenith=None, emin=1, emax=100, **integargs):
+    def integrate(self, zenith=None, emin=1, emax=100, interpolate=True, **integargs):
+        logger.debug("Interpolate? {}".format(interpolate))
         return romberg(self, emin, emax, vec_func=True, **integargs)
 
     def integrate_samples(self, energy, zenith=None, emin=1, emax=100,
-                          **integargs):
+                          interpolate=True, **integargs):
+        logger.debug("Interpolate? {}".format(interpolate))
         energy = np.atleast_1d(energy)
         mask = (emin <= energy) & (energy <= emax)
         energy = energy[mask]
         if zenith:
+            logger.debug("Zenith available, using angle-dependent table...")
             zenith = np.atleast_1d(zenith)
             zenith = zenith[mask]
-        flux = self(energy, zenith=zenith)
+        flux = self(energy, zenith=zenith, interpolate=interpolate)
         return simps(flux, energy, **integargs)
 
 
@@ -86,7 +99,7 @@ class PowerlawFlux(BaseFlux):
         self.gamma = gamma
         self.scale = scale
 
-    def _averaged(self, energy):
+    def _averaged(self, energy, interpolate=True):
         return self.scale * np.power(energy, -1 * self.gamma)
 
     def integrate(self, zenith=None, emin=1, emax=100, **integargs):
@@ -107,6 +120,19 @@ class Honda2015(BaseFlux):
     Flux table downloaded from http://www.icrr.u-tokyo.ac.jp/~mhonda/
 
     Flux at Frejus site, no mountain, solar minimum.
+
+    Units are (1 / m^2 sec sr GeV)
+
+    Methods
+    =======
+    __init__(flavor='nu_mu')
+        Load flux table for the given neutrino flavor.
+    __call__(energy, zenith=None)
+        Return the flux on energy, optionally on zenith.
+    integrate(zenith=None, emin=1, emax=100, **integargs)
+        Integrate the flux via omberg integration.
+    integrate_samples(energy, zenith=None, emin=1, emax=100)
+        Integrate the flux from given samples, via simpson integration.
     """
     def __init__(self, flavor='nu_mu'):
         self.table = None
@@ -138,8 +164,9 @@ class Honda2015(BaseFlux):
             self.table
         )
 
-    def _averaged(self, energy, interpol=True):
-        if not interpol:
+    def _averaged(self, energy, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
+        if not interpolate:
             fluxtable = self.avtable
             ene_bin = np.digitize(energy, self.energy_bins)
             ene_bin = ene_bin - 1
@@ -148,11 +175,12 @@ class Honda2015(BaseFlux):
             flux = splev(energy, self.avinterpol)
             return flux
 
-    def _with_zenith(self, energy, zenith, interpol=True):
+    def _with_zenith(self, energy, zenith, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
         energy = np.atleast_1d(energy)
         zenith = np.atleast_1d(zenith)
         cos_zen = np.cos(zenith)
-        if not interpol:
+        if not interpolate:
             fluxtable = self.table
             ene_bin = np.digitize(energy, self.energy_bins)
             zen_bin = np.digitize(cos_zen, self.cos_zen_bins)
@@ -181,10 +209,10 @@ class HondaSarcevic(BaseFlux):
         # adjust upper bin for the case zenith==0
         self.cos_zen_bins[-1] += 0.00001
 
-    def _averaged(self, energy):
+    def _averaged(self, energy, interpolate=True):
         raise NotImplementedError("Supports only zenith dependent flux!")
 
-    def _with_zenith(self, energy, zenith):
+    def _with_zenith(self, energy, zenith, interpolate=True):
         fluxtable = self.table
         cos_zen = np.cos(zenith)
         ene_bin = np.digitize(energy, self.energy_bins)
@@ -196,87 +224,98 @@ class HondaSarcevic(BaseFlux):
 
 class DarkMatterFlux(BaseFlux):
     """
-    Get Dark Matter spectra (taken from M. Cirelli).
+    Get Dark Matter WimpWimp->foo->nu spectra.
+
+    Units are ``d/d n_particles / cm^2 sec sr``.
 
     >>> from km3flux import DarkMatterFlux
     >>> flux = DarkMatterFlux('anu_mu')
-
-    >>> ene = np.logspace(0, 2, 11)
-
     >>> flux(ene)
     array([  6.68440000e+01,   1.83370000e+01,   4.96390000e+00,
          1.61780000e+00,   5.05350000e-01,   2.29920000e-01,
          2.34160000e-02,   2.99460000e-03,   3.77690000e-04,
          6.87310000e-05,   1.42550000e-05])
 
+    Methods
+    =======
+    __init__(source='gc', flavor='nu_mu', channel='w', mass=1000)
+        Load flux table for the given neutrino flavor.
+    __call__(energy, zenith=None, interpolate=True)
+        Return the flux on energy, optionally on zenith.
+    integrate(zenith=None, emin=1, emax=100, **integargs)
+        Integrate the flux via romberg integration.
+    integrate_samples(energy, zenith=None, emin=1, emax=100)
+        Integrate the flux from given samples, via simpson integration.
+
+    Example
+    =======
+    >>> from km3flux import DarkMatterFlux
+    >>> print(DarkMatterFlux.flavors)
+    >>> flux = DarkMatterFlux(flavor='anu_mu', channel='w', mass=1000)
+    >>> ene = np.geomspace(1, 100, 11)
+    >>> print(flux(ene, interpolate=True))
     """
+    flavors = DM_GC_FLAVORS
+    channels = DM_GC_CHANNELS
+    masses = DM_GC_MASSES
+
     def __init__(self, source='gc', flavor='nu_mu', channel='w', mass=1000):
-        self.flavor = flavor
-        self.channel = channel
-        self.mass = mass
-        if source == 'sun':
-            loader = dm_sun_spectrum
+        """
+        Parameters
+        ==========
+        source: string, optional [default: 'gc']
+            Object where the WimpWimp annihilates. Currently only 'gc'
+        flavor: string, optional [default: 'nu_mu']
+            Neutrino flavor. See available flavors stored in
+            ``DarkMatterFlux.flavors``
+        channel: string, optional [default: 'w']
+            WimpWimp annihilation channel (WimpWimp -> foofoo).
+            See available channels stored in ``DarkMatterFlux.channels``
+        mass: int, optional [default: 1000]
+            WimpWimp parent mass.
+            See available masses stored in ``DarkMatterFlux.masses``
+        """
+        if source == 'gc':
+            fname = DM_GC_FILE
         else:
-            loader = dm_gc_spectrum
+            raise NotImplementedError('Only GC supported so far!')
+        self.mass = mass
+        self.channel = channel
+        self.flavor = flavor
 
-        self.counts, self.lims = loader(flavor=flavor, channel=channel,
-                                        mass=mass, full_lims=True)
+        tab = pd.read_hdf(fname)
+        tab = tab[(tab.flavor == flavor) & (tab.mass_dm == mass)]
+        self.flux = tab[channel].values
+        self.x_energy = tab['energy']
+        self.avinterpol = splrep(
+            self.x_energy,
+            self.flux,
+        )
 
-    def _averaged(self, energy):
+    @property
+    def points(self):
+        return self.x_energy, self.flux
+
+    def _check_energy(self, ene, epsilon=0.0001):
+        if np.any(ene > self.mass + epsilon):
+            raise ValueError('Energies exceed parent mass!')
+        return ene
+
+    def __call__(self, energy, interpolate=True):
         energy = np.atleast_1d(energy)
-        if np.any(energy > self.lims.max()):
-            raise ValueError(
-                "Some energies exceed parent mass '{}'!".format(self.mass))
-        fluxtable = self.counts
-        ene_bin = np.digitize(energy, self.lims)
-        ene_bin = ene_bin - 1
-        return fluxtable[ene_bin]
+        energy = self._check_energy(energy)
+        flux = splev(energy, self.avinterpol)
+        return flux
 
-    def _with_zenith(self, energy, zenith):
+    def _averaged(self, *args, **kwargs):
+        return self(*args, **kwargs)
+
+    def _with_zenith(self, energy, zenith, interpolate=True):
+        logger.debug("Interpolate? {}".format(interpolate))
         logging.warning('No zenith dependent flux implemented! '
                         'Falling back to averaged flux.'
                        )
         return self._averaged(self, energy)
-
-
-def dmflux(energy, **kwargs):
-    """Get the DM flux for your energies."""
-    dmf = DarkMatterFlux(**kwargs)
-    return dmf(energy)
-
-
-def e2flux(energy, **kwargs):
-    pf = PowerlawFlux(**kwargs)
-    return pf(energy)
-
-
-def all_dmfluxes(source='gc'):
-    """Get all dark matter fluxes from all channels, masses, flavors."""
-    fluxes = {}
-    if source == 'sun':
-        flavors = DM_SUN_FLAVORS
-        channels = DM_SUN_CHANNELS
-        masses = DM_SUN_MASSES
-    else:
-        flavors = DM_GC_FLAVORS
-        channels = DM_GC_CHANNELS
-        masses = DM_GC_MASSES
-    for flav in flavors:
-        for chan in channels:
-            for mass in masses:
-                mass = int(mass)
-                fluxname = (flav, chan, mass)
-                fluxes[fluxname] = DarkMatterFlux(source=source, flavor=flav,
-                                                  channel=chan, mass=mass)
-    return fluxes
-
-
-def all_dmfluxes_sampled(energy, **kwargs):
-    fluxes = all_dmfluxes(**kwargs)
-    sampled_fluxes = {}
-    for fluxname, flux in fluxes.items():
-        sampled_fluxes[fluxname] = flux(energy)
-    return fluxes
 
 
 class AllFlavorFlux():
@@ -286,7 +325,7 @@ class AllFlavorFlux():
     =======
     __init__(fluxclass='Honda2015')
 
-    __call__(energy, zenith=None)
+    __call__(energy, zenith=None, interpolate=True)
         Return the flux on energy, optionally on zenith.
     """
     fluxmodels = {
@@ -301,7 +340,8 @@ class AllFlavorFlux():
         for flav in FLAVORS:
             self.flux_flavors[flav] = fluxclass(flav)
 
-    def __call__(self, energy, zenith=None, flavor=None, mctype=None):
+    def __call__(self, energy, zenith=None, flavor=None, mctype=None,
+                 interpolate=True):
         """mctype is ignored if flavor is passed as arg."""
         if mctype is None and flavor is None:
             raise ValueError("Specify either mctype(int) or flavor(string)")
@@ -317,6 +357,7 @@ class AllFlavorFlux():
             where = flavor == flav
             if zenith is not None:
                 loczen = zenith[where]
-            flux = self.flux_flavors[flav](energy[where], loczen)
+            flux = self.flux_flavors[flav](energy[where], loczen,
+                                           interpolate=interpolate)
             out[where] = flux
         return out